Labs: Use FW_RSC_ADDR_ANY for vring device addresses
The current examples use a value of 0 for da in the vrings in the
resource table, with the Linux kernel filling in the actual allocated
address. The newer Linux kernels (> 4.19) update this vring da field
only if it is set as FW_RSC_ADDR_ANY. So, update the resource tables
to use this value instead of 0 to get these firmwares to work with
the newer kernels. The change is backward compatible for existing
kernels.

Signed-off-by: Nick Saulnier <nsaulnier@ti.com>